Add fixed base file
