+++ plonky2/src/gadgets/mod.rs
@@ -3,9 +3,9 @@ pub mod arithmetic_extension;
 pub mod arithmetic_u32;
 pub mod biguint;
 pub mod curve;
-pub mod curve_windowed_mul;
-// pub mod curve_msm;
+pub mod curve_fixed_base;
 pub mod curve_msm;
+pub mod curve_windowed_mul;
 pub mod ecdsa;
 pub mod glv;
 pub mod hash;
